Form able to add an element (debug needed)
I'm satsified with the logic and methods of
the jQuery program, which is supposed to make
a form that interactively expands / contracts
to enter data in a MySQL database. Trying to add
form fields on the fly.

I'm happy with the logic and methods of the program
But currently it's completely broken and needs
debugging. I'm not sure how to debug jQuery right
now