Dynamically generate form
Previously, habit-form.php only brought forth a list of habits to
the screen. Now it generates an entire form for interacting with
the database. Habits are labeled, and have fields for indicating
completion of the field and whether it should have more priority.
Also has a crude score calculation.

The form as it stands is fine. What's missing now is the actual logic
that will make the application usable.