Ajax complete. Form sends info to server
In my habit program, I wanted to make a one page application, Each
click sends data to the system (whether the habit is not a priority
or completed). When an item is clicked, it disappears until the
page is refreshed. Ajax sends the info a habit-complete.php for
further processing. The info isn't sent to a database and adequate
logic has yet to be prepared.

Also, a bit of spring cleaning in some other files.