Habit complete now updates habit database
Here is a basically working php script that sends can receive the
data it gets from the dynamic habit form and put it into a database.
The one thing missing is to work out the not intuitive way that
checkboxes get serialized into post variables will the form is
being manipulated by an active script.