Verbosity: Habit-complete.php returns JSON
Currently program development is stalled. It isn't clear to me,
as a user of my own program, how my habits are being tracked in
a manner that's useful to me. This new code outputs ajax formatted
in the habit-complete.php file and puts it directly into an unordered
list item