Key Constraints Added To Database
During use of the program I noticed it urging me onto to one habit
out of all proportion to how important it was. This is because I
had a bad data point that would have been avoided with a primary key
constraint. I also added foreign keys to keep give the database
referential integrity.