Proposed Code for Json Manipulation
An outline of how my program will pick up json from the habit.json
file, turn it into an associate array,  and manipulate it.
Currently, my programming is not behaving as it should, even though
the tests are passing. Rewrite the tests so that they test what you predict.

In this case, your monthly and weekly tests are decrementing values and
the should be incrementing them.