Habits now marked visually via color
Instead of a sorting system as I've thought of previously, I no
have a habit tracker that is color coated. I begins with black,
grows greener and eventually hits dark red. Due to the nature
of color, this leads my logic to be based on 256. Not 7 or 30,
as would be the case with weeks or months. Habit-cron has yet
to be updated to reflect this logic.