Layout Logic, Classes, and New Unit Tests
I've decided to make my layouts for habits fine grained, and
illustrated through border gradients. As a habit gets closer
to being necessary, a table row with a gradient background will
appear. The gradient going from black / dark green towards red.

I have unit tests and classes to write. In this case, a habit is
a class. It will set it's urgency, and describe what the CSS
response should be. My unit tests should show my habit urgency
algorithms are passing in a certain range.

I also have habits for when cron jobs should be called. To be
honest, cron jobs only run if a computer is turned on all the
time. My program should be able to calculate where I stand at
run time, when the page is dynamically generated. Should
remove this file at a later commit